+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   `InputRichBlockAnimation`, `InputRichBlockAudio`, `InputRichBlockPhoto`, `InputRichBlockVideo`,
   `InputRichBlockVoiceNote`, `InputRichBlockThinking`.
 - New #211: Add `isEphemeral` field to `BotCommand` type.
+- New #211: Add `receiverUser` field to `Message` type.
 - Enh #211: `RichText` types, `RichBlockCaption`, `RichBlockTableCell` and `Location` now support serialization
   back into a request array via `toRequestArray()`.